fix: refactor and make block functions reusable
